requirements.txt
pygithub==1.54.1
python-dotenv==0.15.0
requests==2.25.0
tabulate==0.8.9
